+++ lib/database/lib/model/backup-summary.ts
@@ -115,6 +115,8 @@ export namespace BackupSummary {
     Deleted = "DELETED",
     Failed = "FAILED",
     Restoring = "RESTORING",
+    Canceling = "CANCELING",
+    Canceled = "CANCELED",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.