+++ lib/database/lib/model/backup.ts
@@ -108,6 +108,8 @@ export namespace Backup {
     Deleted = "DELETED",
     Failed = "FAILED",
     Restoring = "RESTORING",
+    Canceling = "CANCELING",
+    Canceled = "CANCELED",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.