+++ lib/oda/lib/model/parameter-definition.ts
@@ -74,6 +74,7 @@ export namespace ParameterDefinition {
     Uri = "URI",
     Url = "URL",
     Number = "NUMBER",
+    Boolean = "BOOLEAN",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.